refactor : move file UserInputValidator to util package
