+++ src/test/java/calculator/util/UserInputValidatorTest.java
@@ -1,8 +1,9 @@
-package calculator.domain;
+package calculator.util;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 import calculator.constants.Constants;
+import calculator.domain.StringCalculator;
 import org.junit.jupiter.api.Test;
 
 public class UserInputValidatorTest {
